requirements-qr.txt
qrcode[pil]>=7.4.2,<9
Pillow>=10.0.0,<12
